Cope with null values in JSON sources
Once the RML-Mapper encounters a json property with value `null`
it stops processing that property for all objects.
To combat this we detect null values while processing and tell the
processor the value is an empty string `""`.
This results in that specific occurrence of the property not being
triplified, which is desired behavior.